jobs/bump-lockfile: fix AWS credential usage for remote builders
For remote builders we need to forward the credential to the remote
and have it loaded that way. It turns out we have a helper for this
in our utils.groovy library so let's just use it. Fixup for d5db023.

Without this we see an error like the following for the remote builders:

```
[2023-01-24T20:52:22.844Z] botocore.exceptions.NoCredentialsError:
Unable to locate credentials
```