jobs/bump-lockfile: lower number of artifacts to keep
This job was taking up 7G of storage on the PVC in the FCOS pipeline.
Since it runs tests on all architectures it will naturally take up way
more storage as it archives all of those kola logs. Let's lower the
number of artifacts it stores to 30 to alleviate this storage usage.