config: add `streams[].skip_disk_images`
As part of https://github.com/openshift/enhancements/pull/1637, we've
moved OCP 4.19 to use bootimages with RHEL content only. This means that
the bootimages built with OCP content will never be used in practice.

Add a `skip_disk_images` knob to disable building them.

We still generate the QEMU image for kola tests to run and because
they're useful to debug, but we drop everything else. (Actually, we
could also not generate the QEMU image either and sanity-check the OCI
image with `kola run --oscontainer`, but that requires more rewiring.)

We don't generate live media since I don't think the test coverage from
that is meaningfully different enough from the RHEL-only variants given
that the additional OCP packages are unrelated.