config: add `streams[].skip_cloud_images`
In RHCOS, we're often in a state where we don't care about regenerating
all the disk images and instead just want the containers. We could
do this by manually adding all the images we don't care about as
`skip_artifacts` entries, but that's cumbersome and error-prone.

Add a `skip_cloud_images` knob for this. We still generate the QEMU
and live artifacts for kola tests to run and because they're useful to
debug, but we drop everything else.