jobs/build: factor out functions to upload OSTree and trigger jobs
Factor out a function to trigger the release job since we'll be calling
it from another location in a future patch.

Transform the closures for uploading the OSTree and triggering the
multi-arch jobs into separate functions. Having them as closures makes
it look more complex than it really is.

While we're here, consistently wrap the stages inside their gating
if conditions rather than the other way around. We already did this for
the "Publish" stage but not for the other stages involved here.

We could pass less variables if we make `basearch` and `newBuildID`
script globals, but because they're mutable, I'd like to avoid that to
keep it easier to follow.

While we're here, update the outdated comment near where we trigger the
release job.