jobs/build: add `WAIT_FOR_RELEASE_JOB` parameter
In RHCOS, the pipeline is mainly driven by higher-level tools that
orchestrate a larger pipeline. In that context, it's not helpful to have
the `build` job just trigger the `build-arch` jobs and `release` jobs
and exit. Automation would have to independently track the state of the
triggered jobs and wait for them. It's much simpler to be able to wait
for just the `build` job.

This patch adds a new `WAIT_FOR_RELEASE_JOB` parameter which makes the
`build` job wait for the `release` job to finish, propagating failures
accordingly.