jobs/build: rerun `build-arch` if previous build is incomplete
Currently, if the x86_64 build succeeds, but not e.g. the aarch64 build,
a rerun of the `build` job will no-op. Salvaging the build requires
rerunning just the `build-arch` job for aarch64, and then rerunning the
`release` job. This is relatively easy for a human, but isn't very
automation-friendly.

With this patch, if we no-op but the latest build is missing builds for
some of the architectures, we automatically rerun the `build-arch` job
for those missing arches and then rerun the `release` job. This allows
the interface for automation to solely be the `build` job, and allows
transparently salvaging up-to-date but incomplete builds rather than
paying for a much costlier full rebuild on all arches.