jobs/build: increase timeout when WAIT_FOR_RELEASE_JOB set
Our default 240 minute timeout won't be enough when
`WAIT_FOR_RELEASE_JOB` is set since the release job itself needs to wait
for all the build-arch jobs to finish before proceeding. Set the timeout
to a higher value accordingly.

A downside with this is that it increases the timeout for all the other
bits of the job too. Notably cloud uploads are susceptible to hangs if
the infra is flaky (as it recently happened with Alibaba). To counter
this, add a timeout to cloud uploads. I chose 45 mins after looking at
past times in the available build history; the longest cloud uploads
ever took was 23 mins, so 45 mins should be a safe value. We can add
more scopes elsewhere in the future if we feel it's warranted.