jobs/build: wait until build-arch takes lock after starting
This solves a potential race between triggering the build-arch and
releaes job where the latter may take the release locks before the
build-arch does. This wasn't likely before (though still theoretically
possible) since we did many things between triggering the two jobs. But
now in the new "complete the previous build" path, they're triggered one
after the other so the risk is much higher.

The technique here isn't foolproof. If the job fails early on (e.g. `git
clone` failure), we'll sit there waiting for something that'll never
happen. To counter this, add a timeout. We don't make it fatal because
we still want the semantic of a "best-effort release" to apply.