jobs: make pipeutils, pipecfg, and libcloud script globals
In Groovy, variables declared at the topmost scope of the script with
`def` are accessible to the rest of the script and inner scopes. But any
function declared in the script is considered a separate scope and will
not have access to these variables.

I'd like to introduce functions and have those functions have access to
`pipeutils` and `pipecfg`. Since these are generic "immutable" (though
not enforced) variables, let's drop the `def` for them so that they're
"script global".