config: support per-stream `skip_cloud_uploads` knob
While working with unreleased content, we don't want any
publicly accessible artifacts. This includes cloud images. Add a
`skip_cloud_uploads` key for this.

This is temporary to unblock the RHEL 9.2 work. I think what would be
nicer here instead is to still upload to clouds but keep the images
private. However, we need to add a `--public` knob to some of the cloud
commands we use first.

Another option worth investigating is support for overriding the base
`cloud` key entirely like we have for some other keys.