jobs: add kola-upgrade test
This will test upgrades from older releases to verify old systems
can be updated to the latest versions. For our production FCOS streams
we'll kick off a full matrix to try to test from different points in
the history of the stream. For development/mechanical streams we'll
just kick of a single modulo $date test per build.