jobs:debug-pod for multi-arch builder access
In order to resolve the limited access to multi-arch for most of our
team members,we have the multi-arch builders where we could just use
the pipeline itself to spawn a remote session on a builder and the user
can then grab a shell via OpenShift to do testing. cosa-init & sleep
stages are added as part of creating a job for multi-arch builder access.

Co-authored-by: Dusty Mabe <dusty@dustymabe.com>
Ref: https://github.com/coreos/fedora-coreos-pipeline/issues/803