jobs/kola-upgrade: use latest in `start_stream` if necessary
We hit this today in the pipeline. The `builds.size() % max-min+1`
code specified 39 as the start_version (rawhide is currently on 39)
but the `next` stream is on 38. So the `releases.json` traversal code
wasn't finding any builds in the `next` stream and the test ultimately
failed.

If we encounter a case like this, let's just use the latest build
from the `start_stream`.