config: add container image tags to push to and drop add_build_tags
This commit updates the config to be able to specify which tags to
push to inside the registry_repos definition. For example, we went
from:

  oscontainer: quay.io/fedora/fedora-coreos

to:

  oscontainer:
    repo: quay.io/fedora/fedora-coreos
    tags: ["${STREAM}"]

This allows us to do things like embed any needed "tag_suffix" in
the config rather than in the code. i.e. downstream now our extensions
container entry will look like:

  extensions:
    repo: quay.io/openshift-release-dev/ocp-v4.0-art-dev
    tags: ["${STREAM}-extensions", "${VERSION}-extensions"]

It also means we can now get rid of the `add_build_tag` hack that
we had in the past.