jobs/kola-{openstack,upgrade): adjust to coreos-ci-lib allowRerunSuccess changes
In https://github.com/coreos/coreos-ci-lib/pull/148
we are modifying the defaults to now call kola run with
`--allow-rerun-success tags=needs-internet`, which is part of
https://github.com/coreos/fedora-coreos-pipeline/issues/842.

Here we drop the `--allow-rerun-success tags=all` call from the upgrade
test. Since the test itself has the `needs-internet` tag then it will
be covered by the new default.

For the openstack test we'll adjust to the new way to override the
argument string passed to `--allow-rerun-success` inside of `kola()`.