Revert "release: Merge brew job with release job"
When these two jobs were merged, the secrets for RHCOS are now being
included when creating the pod in FCOS. This causes the pod to fail.
Revert this change to unblock the prod stream releases while we
determine how to differentiate the two.

reverts: 5d92f139fc27e9db28714464c6dddbcdef535236