manifests/jenkins-s2i: set `scheduled` on our imagestreams
I'm not sure why we didn't do this before. Without this, the
imagestreams are only updated manually. In the `jenkins-with-cert`
flow, we already have an `ImageChange` trigger which ensures that
`jenkins:latest` is kept fresh. So this change is putting the non-
custom-cert path at parity.

Note that this is separate from starting and deploying the S2I build,
which currently is still done manually.