jobs/build-arch: only run one ppc64le build at a time with ppc64le_kola_minimal
Our builder only has 32G of RAM. Combined with the fact that ppc64le
VMs require a minimum of 4G of RAM, that doesn't leave much capacity for
parallelism. We don't run tests, but multiple artifacts launch VMs.

Let's bring it down to 1 build at a time to make it more reliable.
If the throughput is too slow, we'll bump to a beefier instance and
increase parallelism here.