utils: fix azure launching login in run_cloud_tests()
In 582bfae we changed the azure creds used for testing to a single
file/secret. What we didn't do was update the run_cloud_tests() logic
to check for the new secret name, it was still using the old secret
names.

I assume this continued to work because the old secrets still existed
in the pipeline until I recently deleted them because our new secret
expired and I noticed the old named ones existed and deleted them.

Fixes 582bfae