config: promote `versionary` key, rework it
The versionary FCOS script was one of the few things that differed
between FCOS and RHCOS and was awkwardly attached as a hack knob to
work.

Now, we promote it to a proper key under the `misc` section, with
support for stream-level overrides. This is prep for also using a
versionary in other streams.

The versionary is now part of the source config and cosa knows how to
execute it if the `--versionary` argument is passed.