jobs/build-development: make development streams build daily
Over time I'm thinking the build on every push to the git repo is
a bit heavy. Often times the changes will be insignificant to warrant
building, testing, and pushing all that we do. On the other hand, if
there is a change we want we can easily click a button and start a
build if we don't want to wait.