jobs/build-development: make development streams build daily
Over time I'm thinking the build on every push to the git repo is
a bit heavy. Often times the changes will be insignificant to warrant
building, testing, and pushing all that we do. On the other hand, if
there is a change we want we can easily click a button and start a
build if we don't want to wait.

This commit switches the development streams to build daily similarly
to the mechanical ones (i.e. rawhide). It also changes the timing of
the build-mechanical and bump-lockfiles jobs and adds a stage to the
bump-lockfile job to trigger a build once it has pushed.

The timings are chosen based on:

- bump-lockfiles at 04:00 UTC
    - The bodhi/pungi updates composes seem to run around 0:00 UTC
      and finish within an hour or two. This gives some extra time
      so we hopefully trigger a bump-lockfile with the latest content.
- build-development at 05:00 UTC
    - Triggered soon after the bump-lockfiles job it will now wait
      if there are bump-lockfile jobs running before kicking off the
      development stream builds.
    - If the bump-lockfile runs failed (and thus didn't trigger
      a new build at the end of the run) this daily trigger will
      make sure we still get a build. If the bump-lockfile run
      finished and a new build was started then this build will
      wait and hopefully result in "No New Build".
- build-mechanical at 10:00 UTC
    - The rawhide composes seem to run around 5:00 UTC and finish
      around 7:00 UTC or 8:00 UTC. This runs a rawhide soon after
      with an hour or so of grace in case the pungi runs start
      taking more time.