jobs/sync-stream-metadata: send Slack message on failure
This unassuming job is actually critical to the release pipeline flow.
We have checkbox items in the release checklist to make sure that it
ran correctly, but for other things (e.g. reverting a rollout, or adding
release notes), there's zero visibility into whether the sync actually
happened unless one goes to check the job.

Let's just always send a Slack notification if the job fails.

Prompted by noticing that it failed recently because of
https://github.com/coreos/fedora-coreos-streams/pull/924 only by pure
chance (I was looking at the job for another reason).

Patch best viewed with whitespace ignored.