jobs/cloud-replicate: make aliyun idemopotent
Previously when the cloud-replicate job failed replicating on Aliyun
re-running the job would fail with a duplicate image error. The job
would exit early and not execute the meta.json upload. We need to
always upload the meta.json so the replication progress can be
preserved for subsequent runs and avoid the duplicate image error.