jobs/kola-upgrade: account for overhead in x86_64 case
When requesting memory for a COSA pod for x86_64 in the past
(274a97a) we basically requested a pod with exactly the amount of
memory we needed for the 3 virtual machines (bios,uefi,uefi-secure).
This lead to a VM getting killed occasionally at first and then
more consistently over time [1]. Most often this would be the
uefi-secure VM most likely because that process was using more
memory than the other two.

[1] https://github.com/coreos/fedora-coreos-tracker/issues/1756