manifests: use scheduled-upgrade as source imagestream tag
The `jenkins:2` tag hasn't been rebuilt in almost a year and the Jenkins
version and plugin versions are getting out of date. The `scheduled-upgrade`
and `scheduled-upgrade-redeploy` tags are rebuilt when there are changes
in openshift/jenkins. By using these tags instead, we can keep our Jenkins
insance more up-to-date. The `scheduled-upgrade-redeploy` tag will even
automatically redeploy Jenkins when new changes are detected. This has
been working in the production RHCOS pipeline for a while now.