muliti-arch-builders/splunk: only build the splunk sidecar container
The repo we are using to build the ITPAAS splunkforwarder container
on the RHCOS multi-arch builders is out of date and references a
splunkforwarder rpm that no longer exists in the internal repos.
By omitting the build step for this container, we can pull in the
already built image from ITPAAS that includes a more recent RPM.
We'll still use `--arch=x86_64` for the sidecar container to ensure
proper logging on non x86_64 builders.

See: https://github.com/coreos/fedora-coreos-pipeline/issues/1044