jobs: run `cosa push-container-manifest` privileged
In a recent 4.16 z-stream release, a cri-o backport changed the default seccomp
policy to by default block `clone(CLONE_NEW*)` syscalls:

https://github.com/cri-o/cri-o/pull/8514

This affects us in the FCOS pipeline which runs in a cluster that was
recently updated. The `podman manifest` commands all AIUI also flow
through the default path where it wants to enter a namespace if running
rootless even though we don't strictly need root; we're not running
containers, just creating manifest lists. Ideally podman would be less
eager there.

Anyway, work around this as necessary by running
`cosa push-container-manifest` privileged.

There are two general places where this command is used: in container
image build jobs (e.g. `build-cosa`), and in the release job.

In the former, just use one of the multi-arch builders to do this since
we already have a session there.

In the latter, just run it in supermin. This has some warts: we want to
ideally keep uploading in parallel and while it'd be nice to parallelize
*inside* the supermin VM, the tooling doesn't make that easy. Instead,
we run multiple supermin VMs in parallel which means bumping the
resource request.