utils: add new `scheduled_streams` method
`scheduled_streams` takes in a subset of streams found using
`streams_of_type` and returns a list of those streams that have the
new `scheduled: true` config knob set. This is used in the rework
of the build-mechanical job to conditionally build streams with
the `scheduled:true` knob set.

see: https://issues.redhat.com/browse/COS-3012