jobs/build: set EARLY_ARCH_JOBS to false by default
There is a lot of wasted work and leaked dead builds when
EARLY_ARCH_JOBS is set. It has also caused downstream problems
when someone forgets to unset it and incomplete builds are picked
up. Lets turn it off by default and have the user enable when they
need the time savings.