jobs/release: support triggering build-node-image
We want the ability to trigger build-node-image from the release job
instead of the latter being cron triggered. That way we reduce the
chance of having compose skew when in pre-GA state, and we get fresh
node images soon after the base image is updated.

I had initially made the mechanism more generic, but meh, it didn't seem
worth the complexity. We can always generalize it if another "follow-up
trigger" case shows up.