jobs/node-image-build: Use manifest digests instead of tags when pushing
Retrieve the digests of intermediary manifests instead of relying on the
tags they were pushed with.  This prevents accidentally referencing
stale manifests that may still exist locally on the builders from
previous runs.

We were already using digests for the images within the manifests, so
this also improves overall consistency.

Co-authored-by: Renata Ravanelli <rravanel@redhat.com>