jobs/node-image-build: use manifest digests instead of tags for pushing
Retrieve the digests of the intermediary manifests
instead of the tags we pushed them to.
This makes sure we are not accidentally using an
older manifest that is still laying around from a
previous run.

We were already doing that for the actual images
inside the manifests so it's also more consistent.