jobs/build-node-image: label images with manifest list tag
ART needs to be able to, from an x86_64 nightly, derive an s390x version
of it. They used to be able to do this before the layering work by
looking at the build ID, but now the source of truth is the manifest
list, which is a moving target.

Enable this by adding the manifest list tag to all the built images.
This feels like a layering violation of course. I think the more
appropriate fix for this is to better keep track of the manifest lists
used from the start when building arch-specific nightlies, but for now
at least let's not break current workflows. And anyway, all this will
have to be reworked for Konflux.