jobs/build-fcos-buildroot: don't read pipecfg
While having the buildroot image be defined in fedora-coreos-config is
on purpose, in the context of actually building the image we shouldn't
need the pipecfg here. It's just a given that the canonical image
definition is in f-c-c, so let's hardcode it and drop the pipecfg
sourcing. This matches e.g. `build-cosa`.