libcloud: update aws-winli creation to use billing code
The Windows License Included (winli) AMI creation logic was reworked in
coreos assembler [1] to append the winli billing code to an
already-created RHCOS AMI, instead of creating an AMI based on a
modified Windows instance. Update the winli build closure to use the new
imageupload-aws option. We no longer need to discover and use the latest
windows server 2022 AMI ID.

[1]: https://github.com/coreos/coreos-assembler/pull/4237