Add build-nvidia-bfb job
This job is for a proof-of-concept downstream effort to build `BFB` files
to boot on NVIDIA BlueField DPU hardware [1]. It leverages the `cosa import`
flow to pull a customized RHCOS container and build boot media from it.

This job was initially copied and based on the build-arch job since this
platform is `aarch64` only.

[1] https://docs.nvidia.com/networking/display/bluefielddpuosv450/deploying+bluefield+software+using+bfb+from+host