jobs/release: fix Brew upload conditionals
Commit 5ca628f ("jobs/release: only fetch
coreos-assembler-config-git.json if uploading to Brew") would work for
RHCOS in the buildah path, but not FCOS, where we don't even need to
specify `skip_brew_upload` since we don't do any uploads at all and thus
don't have a Koji profile.

Just rework this logic to use the exact same conditional that brew
uploading is gated on.