Revert "jobs/build-arch: add --autolock to the `cosa build` call"
It turns out the underlying code will autolock anyway even without
--autolock if it detects the right conditions so this change didn't
actually do anything. The reason for the inconsistent behavior I was
seeing that caused me to implement this change was fixed by
https://github.com/coreos/fedora-coreos-config/pull/3705

The other reason to revert this is older COSA (i.e. 4.14 and older)
don't recognize this option at all.

This reverts commit 0e1c9742351ec83e1c9978a4cec209997f06c75e.