jobs/release: sign OCI images
This uses the new `cosa sign --oci` switch to sign our OCI images using
the official Fedora GPG keys.

Part of https://github.com/coreos/fedora-coreos-tracker/issues/1969.

I opted for not gating this behind a knob. It's safe to do this even on
streams that still use encapsulated commits (it's just files in S3). But
the earlier we do it, the more overlap we have in which both signature
types (OCI and OSTree) co-exist and that makes migration planning
easier.