jobs/release: simplify pushed image signing
Now that the manifest list is part of `meta.json`, we no longer need to
use `--write-digest-to-file` and `--manifest-list-digest`. This fixes a
subtle issue: if `cosa push-container-manifest` no-ops (because the
images were already pushed), no digestfile is written, and so signing
would fail.

By having it be in `meta.json`, we make it robust against re-running the
release job, which is designed to be idempotent. (Though worth noting
that the signing operation itself is not yet idempotent; we always
re-sign, overwriting any old signature.)

But since signing can commonly fail, let's move it to after we upload
the updated `meta.json` to S3 so that `cosa push-container-manifest` can
at least keep no-op'ing when we rerun.