build: errors if the imported OCI stream mismatch
We need consistency between the imported image stream and what the
pipeline uses. We could silently pull the value from the image but I'd
rather be explicit and fail loudly than create confusion.

In the future we may drop the STREAM parameter and rely entirely on the
image tag but right now we need to support both.

As is, this change won't work with RHCOS images because the key for this
value is `fedora-coreos.stream` which is a bit weird IMO. It looks like
this is legacy from the ostree commit metadata that was carried over.
Although we aren't going to turn this on (importing konflux images) for
RHCOS soon it's still worth thinking about it I think.

Maybe we could change it to `com.coreos.stream` to match the other label
(`com.coreos.osname`) we have ?

This would requires changes around other places, e.g Zincati [1] and cosa.

See https://github.com/coreos/fedora-coreos-pipeline/pull/1185#discussion_r2342187389