Add locking to make kola-upgrade jobs run after release job.
The release job is the one that pushes (and signs) containers.
If we want our kola-upgrade test to be more realistic it makes
sense to have it run after the containers have been pushed so we
can test pulling from the registry and also signature verification.

Here we make the kola-upgrade job wait until the release job is done
before starting the upgrade tests. For the production builds this
won't really have any effect because we don't autorun the release
job, but that's OK.