multi-arch-builders: document us-south region for s390x
The ca-tor region has IP addresses that geoip databases think are
in France so performance in that region can be un-ideal as it might
try to pull from mirrors in Europe. Let's just document our second
region as us-south.

We've been asked to move out of us-east for now to ease some resource
pressure there. So we're moving to us-south.