Allow for skipping build of untested artifacts
While we are proud of our CI there are many platforms we still don't
do automated testing for. Let's make our mechanical and development
stream builds a little more lightweight by limiting the build and
upload of all of the artifacts we don't automatically test to just
a single day of the week.

This is achieved by adding a new 'SKIP_UNTESTED_ARTIFACTS' parameter
which defaults to `false` (current behavior), but which is set when
the build-deveopment and build-mechanical jobs are run daily.

Resolves https://github.com/coreos/fedora-coreos-pipeline/issues/1189