jobs/bump-lockfile: force fetch to be effective
In the new build-with-buildah path for building `cosa fetch` is
typically useless so we effectively made it exit early [1] [2] if
building with buildah is enabled.

In the case of the bump-lockfile job, though, we are actually using
the fetch to update the lockfiles first, and then use them as inputs
to the build. In this case we need `cosa fetch` to actually update
the lockfiles, so let's override with COSA_BUILD_WITH_BUILDAH=0 here.

[1] https://github.com/coreos/coreos-assembler/commit/0ce8501ded7e3ab4d370e8333b2800716d20d655
[2] https://github.com/coreos/coreos-assembler/commit/bb7443a6c22d336a92b88759417c0a34aa631260