jobs: compute SKIP_UNTESTED_ARTIFACTS default in build
Since the migration to Konflux for rawhide stream, we need to compute
`SKIP_UNTESTED_ARTIFACTS` default value in the `build` job as it's the
one Konflux triggers to run the rest of the pipeline.
We update the other jobs accordingly.
Other option would be to create a dedicated Jenkins job for Konflux
e.g: `build-mechanical-konflux` without the cron part but this brings
more complexity and maintenance work.

Closes [1]

[1] https://github.com/coreos/fedora-coreos-pipeline/issues/1263