build: validate build stream param against OCI image
Adds validation to ensure the build's expected stream, provided by
the `${params.STREAM}` parameter, matches the 'com.coreos.stream'
label found in the imported OCI image.

The OCI image label is treated as the source of truth, and this check
enforces that the build context correctly reflects the image's
intended stream.