build: enforce OCI image import for Konflux-driven streams
We now error (fail the build) when a STREAM driven by Konflux does
not explicitly import an OCI image.

This reverts 0addd8ee60f14b8723ffa8a90a9554da31529e8a. The function
`get_streams_choices` is called in multiple places, so we cannot use
it to limit the scope.