utils: determine day of the week with pure Groovy
Fixes 926658663bd2cd8d0a7febd84a8209fc07e1c83a which executes
`should_we_skip_untested_artifacts` method early in the properties
section which computed the parameters.
However, this method is using a `sh` command but it's failing as it's
not wrapped in a node or agent block.
This patch removes the use of the `shell` and get the day of the week
with pure Groovy instead.

We are using the `java.util.Calendar` API for compatibility with older
Jenkins/Groovy versions.