build: use choice param for SKIP_UNTESTED_ARTIFACTS
We cannot have this method run in parameter scope because the first
build on friday would have `SKIP_UNTESTED_ARTIFACTS` set to True and
the next ones set to False. Jenkins only updates the default value
after the first instanciation of the job.

We move to choice parameter and introduces the dynamic value which
is the default one.

The allowed choices are now explicitly:
- dynamic: rely on the should_we_skip_untested_artifacts() method.
- yes: explicitly set the flag to true.
- no: explicitly set the flag to false.